move content fetching functions to contentHelpers and tidy up a bit
